pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "silicon-design-environment"
version = "0.1.0"
description = "Silicon Design Environment for RISC-V cores"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "prefect==2.14.10",
    "numpy==1.26.2",
    "pandas==2.1.3",
    "matplotlib==3.8.2",
    "jupyterlab==4.0.9",
    "scikit-learn==1.3.2",
    "tqdm==4.66.1",
    "pyyaml==6.0.1",
    "griffe==0.36.1",
    "dataclasses-json==0.6.1",
    "typing-inspect==0.9.0",
    "tomli==2.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    "coverage==7.3.2",
    "pluggy==1.2.0",
    "exceptiongroup==1.1.3",
    "iniconfig==2.0.0",
]

[tool.setuptools]
packages = ["flows"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"
